+++ modules/system/model/SysUserRole.go
@@ -6,8 +6,8 @@ import (
 
 // 用户和角色关联
 type SysUserRole struct {
-	UserId int64 `gorm:"size:20;primary_key;auto_increment;用户ID;" json:"userId"`
-	RoleId int64 `gorm:"size:20;primary_key;auto_increment;角色ID;" json:"roleId"`
+	UserId int `gorm:"size:20;primary_key;auto_increment;用户ID;" json:"userId"`
+	RoleId int `gorm:"size:20;primary_key;auto_increment;角色ID;" json:"roleId"`
 }
 
 func (e *SysUserRole) TableName() string {
